move utility function into separate module
